Securing scripts

+++ convert-a-link-admin.php
@@ -1,7 +1,15 @@
 <?php
-if (! empty($_POST['publisherId'])) {
-    delete_option('cal_publisherId');
-    add_option('cal_publisherId', $_POST['publisherId']);
+if ( ! defined( 'ABSPATH' ) ) exit;
+
+if ( ! empty( $_POST ) && check_admin_referer( 'insert_publisher' ) ) {
+  if (! empty($_POST['publisherId'])) {
+      $publisherId = sanitize_text_field($_POST['publisherId']);
+      $publisherId = intval($publisherId);
+      delete_option('cal_publisherId');
+      add_option('cal_publisherId', $publisherId);
+  } else {
+    echo "<div class='error'>FAILED! Insert publisher ID</div>";
+  }
 }
 ?>
 <div class="wrap" xmlns="http://www.w3.org/1999/html" xmlns="http://www.w3.org/1999/html"
@@ -20,7 +28,8 @@ if (! empty($_POST['publisherId'])) {
             <td><b>Step 2: Enter your publisherId and save.</b></td>
             <td>
                 <form enctype="multipart/form-data" name="convert-a-link" method="post" action="">
-                <input type="number" value="<?=get_option('cal_publisherId')?>" name="publisherId"/>
+                  <?=wp_nonce_field( 'insert_publisher');?>
+                <input type="number" value="<?php echo esc_attr(get_option('cal_publisherId'))?>" name="publisherId"/>
                     <input type="submit" name="submit" id="submit" class="button button-primary" value="Save">
                 </form>
             </td>
